fix(eve): resolve parent package tsconfig path aliases in dev runtime snapshots
A withEve agent whose root is a bare directory inside a Next.js app (no
package.json/tsconfig.json of its own) bundles fine in production because
authored-module bundling walks up to the owning package and passes its
tsconfig path aliases to rolldown. Dev runtime snapshots broke that in two
ways: the snapshot plan only scanned the app root itself for tsconfigs, so
the owning package's tsconfig and alias-target sources were never copied,
and the copy step synthesized a package.json at the snapshot agent root,
moving the package boundary below the owning package so the parent
tsconfig stayed invisible to the bundler.

Mirror the dependency-declaration walk-up (vercel/eve#1151) for tsconfig
discovery: resolve tsconfigs from the nearest owning package as well, copy
that package into the snapshot when its tsconfig declares path aliases,
and only generate a synthetic runtime package.json when no owning
package.json exists at or above the runtime app root inside the snapshot.

Fixes #1242
